Move barcode generation via Baracoda to general helpers
