+++ crawler/exceptions.py
@@ -34,6 +34,13 @@ class DartStateError(Error):
             return f"DartStateError: {default_message}"
 
 
+class BaracodaError(Error):
+    """Raised when communication with Baracoda results in unsuccessful states/outcomes."""
+
+    def __init__(self, message):
+        self.message = message
+
+
 class CherrypickerDataError(Error):
     """Raised during cherrypicker test data creation. The message is assumed to be user friendly."""
 
@@ -42,8 +49,7 @@ class CherrypickerDataError(Error):
 
 
 class TransientRabbitError(Error):
-    """
-    Raised during processing of a RabbitMQ message when a transient issue occurs.
+    """Raised during processing of a RabbitMQ message when a transient issue occurs.
     For example, this might be a database being inaccessible.  The message should be reprocessed.
     """
